Palindromes question
The palindrome detector seems to be a common interview question now.  I didn't finish writing one up on the whiteboard, so I went home and thought about it a bit.  This is as far as I've gotten on it.  Should work for most strings, optionally allows spaces to also be matched, checks in a case-insensitive fashion.